integrate stereo baseline
