+++ src/eval/generate_kitti.py
@@ -8,6 +8,7 @@ import glob
 import json
 import shutil
 import itertools
+import copy
 
 import numpy as np
 import torch
@@ -17,68 +18,148 @@ from eval.geom_baseline import compute_distance
 from utils.kitti import get_calibration
 from utils.pifpaf import preprocess_pif
 from utils.camera import xyz_from_distance, get_keypoints, pixel_to_camera
+from utils.stereo import depth_from_disparity
 
 
-def generate_kitti(model, dir_ann, p_dropout=0.2, n_dropout=0):
+class GenerateKitti:
 
-    cnt_ann = 0
-    cnt_file = 0
-    cnt_no_file = 0
+    def __init__(self, model, dir_ann, p_dropout=0.2, n_dropout=0):
 
-    dir_kk = os.path.join('data', 'kitti', 'calib')
-    dir_out = os.path.join('data', 'kitti', 'monoloco')
+        # Load monoloco
+        use_cuda = torch.cuda.is_available()
+        device = torch.device("cuda" if use_cuda else "cpu")
+        self.monoloco = MonoLoco(model_path=model, device=device, n_dropout=n_dropout, p_dropout=p_dropout)
+        self.dir_out = os.path.join('data', 'kitti', 'monoloco')
+        self.dir_ann = dir_ann
 
-    # Remove the output directory if alreaady exists (avoid residual txt files)
-    if os.path.exists(dir_out):
-        shutil.rmtree(dir_out)
-    os.makedirs(dir_out)
-    print("Created empty output directory for txt files")
+        # List of images
+        self.list_basename = factory_basename(dir_ann)
+        self.dir_kk = os.path.join('data', 'kitti', 'calib')
 
-    # Load monoloco
-    use_cuda = torch.cuda.is_available()
-    device = torch.device("cuda" if use_cuda else "cpu")
-    monoloco = MonoLoco(model_path=model, device=device, n_dropout=n_dropout, p_dropout=p_dropout)
+    def run_mono(self):
+        """Run Monoloco and save txt files for KITTI evaluation"""
 
-    # Run monoloco over the list of images
-    list_basename = factory_basename(dir_ann)
-    for basename in list_basename:
-        path_calib = os.path.join(dir_kk, basename + '.txt')
-        annotations, kk, tt = factory_file(path_calib, dir_ann, basename)
-        boxes, keypoints = preprocess_pif(annotations, im_size=(1242, 374))
+        cnt_ann = 0
+        cnt_file = 0
+        cnt_no_file = 0
+        dir_out = os.path.join('data', 'kitti', 'monoloco_stereo')
+        # Remove the output directory if alreaady exists (avoid residual txt files)
+        if os.path.exists(dir_out):
+            shutil.rmtree(dir_out)
+        os.makedirs(dir_out)
+        print("Created empty output directory for txt files")
 
-        if not keypoints:
-            cnt_no_file += 1
-            continue
-        else:
-            # Run the network and the geometric baseline
-            outputs, varss = monoloco.forward(keypoints, kk)
-            dds_geom = eval_geometric(keypoints, kk, average_y=0.48)
+        # Run monoloco over the list of images
 
-        # Save the file
-        all_outputs = [outputs.detach().cpu(), varss.detach().cpu(), dds_geom]
-        all_inputs = [boxes, keypoints]
-        all_params = [kk, tt]
-        path_txt = os.path.join(dir_out, basename + '.txt')
-        save_txts(path_txt, all_inputs, all_outputs, all_params)
+        for basename in self.list_basename:
+            path_calib = os.path.join(self.dir_kk, basename + '.txt')
+            annotations, kk, tt = factory_file(path_calib, self.dir_ann, basename)
+            boxes, keypoints = preprocess_pif(annotations, im_size=(1242, 374))
 
-        # Update counting
-        cnt_ann += len(boxes)
-        cnt_file += 1
+            if not keypoints:
+                cnt_no_file += 1
+                continue
+            else:
+                # Run the network and the geometric baseline
+                outputs, varss = self.monoloco.forward(keypoints, kk)
+                dds_geom = eval_geometric(keypoints, kk, average_y=0.48)
 
-    # Print statistics
-    print("Saved in {} txt {} annotations. Not found {} images"
-          .format(cnt_file, cnt_ann, cnt_no_file))
+            # Save the file
+            uv_centers = get_keypoints(keypoints, mode='bottom')  # Kitti uses the bottom center to calculate depth
+            xy_centers = pixel_to_camera(uv_centers, kk, 1)
+            outputs = outputs.detach().cpu()
+            zzs = xyz_from_distance(outputs[:, 0:1], xy_centers)[:, 2].tolist()
+            all_outputs = [outputs.detach().cpu(), varss.detach().cpu(), dds_geom, zzs]
+            all_inputs = [boxes, xy_centers]
+            all_params = [kk, tt]
+            path_txt = os.path.join(dir_out, basename + '.txt')
+            save_txts(path_txt, all_inputs, all_outputs, all_params)
+
+            # Update counting
+            cnt_ann += len(boxes)
+            cnt_file += 1
+        print("Saved in {} txt {} annotations. Not found {} images\n".format(cnt_file, cnt_ann, cnt_no_file))
+
+    def run_stereo(self):
+        """Run monoloco on left and right images and alculate disparity if a match is found"""
+
+        cnt_ann = 0
+        cnt_file = 0
+        cnt_no_file = 0
+        cnt_no_stereo = 0
+        cnt_disparity = 0
+        dir_out = os.path.join('data', 'kitti', 'monoloco_stereo')
+
+        # Remove the output directory if alreaady exists (avoid residual txt files)
+        if os.path.exists(dir_out):
+            shutil.rmtree(dir_out)
+        os.makedirs(dir_out)
+        print("Created empty output directory for txt STEREO files")
+
+        for basename in self.list_basename:
+            path_calib = os.path.join(self.dir_kk, basename + '.txt')
+            stereo = True
+
+            for mode in ['left', 'right']:
+                annotations, kk, tt = factory_file(path_calib, self.dir_ann, basename, mode=mode)
+                boxes, keypoints = preprocess_pif(annotations, im_size=(1242, 374))
+
+                if not keypoints and mode == 'left':
+                    cnt_no_file += 1
+                    break
+
+                elif not keypoints and mode == 'right':
+                    stereo = False
+
+                else:
+                    # Run the network and the geometric baseline
+                    outputs, varss = self.monoloco.forward(keypoints, kk)
+                    dds_geom = eval_geometric(keypoints, kk, average_y=0.48)
+
+                    uv_centers = get_keypoints(keypoints, mode='bottom')  # Kitti uses the bottom to calculate depth
+                    xy_centers = pixel_to_camera(uv_centers, kk, 1)
+
+                if mode == 'left':
+                    outputs_l = outputs.detach().cpu()
+                    varss_l = varss.detach().cpu()
+                    zzs_l = xyz_from_distance(outputs_l[:, 0:1], xy_centers)[:, 2].tolist()
+                    kps_l = copy.deepcopy(keypoints)
+                    boxes_l = boxes
+                    xy_centers_l = xy_centers
+                    dds_geom_l = dds_geom
+
+                else:
+                    kps_r = copy.deepcopy(keypoints)
+
+            if stereo:
+                zzs, cnt = depth_from_disparity(zzs_l, kps_l, kps_r)
+                cnt_disparity += cnt
+            else:
+                zzs = zzs_l
+
+            # Save the file
+            all_outputs = [outputs_l, varss_l, dds_geom_l, zzs]
+            all_inputs = [boxes_l, xy_centers_l]
+            all_params = [kk, tt]
+            path_txt = os.path.join(dir_out, basename + '.txt')
+            save_txts(path_txt, all_inputs, all_outputs, all_params)
+
+            # Update counting
+            cnt_ann += len(boxes_l)
+            cnt_file += 1
+
+        # Print statistics
+        print("Saved in {} txt {} annotations. Not found {} images."
+              .format(cnt_file, cnt_ann, cnt_no_file))
+        print("Annotations corrected using stereo: {:.1f}%, not found {} stereo files"
+              .format(cnt_disparity / cnt_ann * 100, cnt_no_stereo))
 
 
 def save_txts(path_txt, all_inputs, all_outputs, all_params):
 
-    outputs, varss, dds_geom = all_outputs[:]
-    uv_boxes, keypoints = all_inputs[:]
+    outputs, varss, dds_geom, zzs = all_outputs[:]
+    uv_boxes, xy_centers = all_inputs[:]
     kk, tt = all_params[:]
-
-    uv_centers = get_keypoints(keypoints, mode='bottom')  # Kitti uses the bottom center to calculate depth
-    xy_centers = pixel_to_camera(uv_centers, kk, 1)
-    zzs = xyz_from_distance(outputs[:, 0:1], xy_centers)[:, 2].tolist()
 
     with open(path_txt, "w+") as ff:
         for idx in range(outputs.shape[0]):
@@ -113,19 +194,25 @@ def factory_basename(dir_ann):
     return list_basename
 
 
-def factory_file(path_calib, dir_ann, basename):
+def factory_file(path_calib, dir_ann, basename, mode='left'):
     """Choose the annotation and the calibration files. Stereo option with ite = 1"""
 
+    assert mode in ('left', 'right')
     p_left, p_right = get_calibration(path_calib)
 
-    kk, tt = p_left[:]
-    path_ann = os.path.join(dir_ann, basename + '.png.pifpaf.json')
+    if mode == 'left':
+        kk, tt = p_left[:]
+        path_ann = os.path.join(dir_ann, basename + '.png.pifpaf.json')
+
+    else:
+        kk, tt = p_right[:]
+        path_ann = os.path.join(dir_ann + '_right', basename + '.png.pifpaf.json')
 
     try:
         with open(path_ann, 'r') as f:
             annotations = json.load(f)
     except FileNotFoundError:
-        annotations = None
+        annotations = []
 
     return annotations, kk, tt
 
@@ -150,4 +237,3 @@ def eval_geometric(keypoints, kk, average_y=0.48):
         dds_geom.append(dd_geom)
 
     return dds_geom
-
